<!-- .github/PULL_REQUEST_TEMPLATE.md -->

## Summary

Diff shows the method used for timing only the ingest path. We simply keep an event counter and print a timestamp when the desired number of events is reached. We skip the `process_cdc_event` call entirely.

Below are the results for both the ingest only path and the full end-to-end run. For each we test and append only insert with 10M rows and 100M rows. 

Results for ingest only path:
`pg` = postres (shows both the insert time for postgres as well as the timestamp)
`ml`= moonlink
`lag`= time between postgres finishing and moonlink finishing

Ingest only:

100M rows: 
```
1) pg: 
INSERT 0 100000000
Time: 124103.400 ms (02:04.103)              
-------------------------------
 2025-08-13 22:05:18.344396+00
 
ml: "2025-08-13T22:05:38.687705125+00:00"

lag: 20


2) pg: 
Time: 159969.330 ms (02:39.969)         
-------------------------------
 2025-08-13 22:25:52.937332+00
(1 row)

ml: 
"2025-08-13T22:25:54.244454917+00:00"

lag: 2 

3) pg: 
Time: 167532.198 ms (02:47.532)             
-------------------------------
 2025-08-14 21:06:51.609207+00
 ml: 
 "2025-08-14T21:06:52.961236843+00:00"
 lag: 1

4) pg: 
Time: 126575.603 ms (02:06.576)             
------------------------------
 2025-08-14 21:11:20.79349+00
 
ml: 
"2025-08-14T21:11:44.748913006+00:00"

lag: 24

5) pg: 
Time: 122273.766 ms (02:02.274)             
-------------------------------
 2025-08-14 21:16:22.247055+00

ml:
"2025-08-14T21:16:45.451574256+00:00"

lag: 23
```

10M rows: 
```
pg: 
Time: 12800.730 ms (00:12.801)    
-------------------------------
 2025-08-14 20:56:49.272404+00
 
 ml: 
 "2025-08-14T20:56:51.717771092+00:00"

lag: 2
 

pg: 
Time: 11543.242 ms (00:11.543)         
-------------------------------
 2025-08-14 20:59:01.637362+00
 
 ml:
 "2025-08-14T20:59:04.646268376+00:00"

lag: 3


pg: 
Time: 11558.610 ms (00:11.559)             
-------------------------------
 2025-08-14 21:00:51.636834+00

ml: 
"2025-08-14T21:00:54.787595177+00:00"

lag: 3
```

End to End: Here we measure the time between the postgres insert and the SELECT * on moonlink

10M

```
pg + select lag: 
Time: 11672.630 ms (00:11.673)
 total_rows 
------------
   10000000
(1 row)

Time: 4333.566 ms (00:04.334)
pg_mooncake (pid: 4068) =# 


pg + select lag: 
Time: 12546.286 ms (00:12.546)
 total_rows 
------------
   10000000
(1 row)

Time: 3947.819 ms (00:03.948)

Time: 12615.459 ms (00:12.615)
 total_rows 
------------
   10000000
(1 row)

Time: 3879.156 ms (00:03.879)
```

100M:

```
pg + lag: 
Time: 149546.324 ms (02:29.546)
 total_rows 
------------
  100000000
(1 row)

Time: 21713.495 ms (00:21.713)

Time: 131960.545 ms (02:11.961)
 total_rows 
------------
  100000000
(1 row)

Time: 34223.637 ms (00:34.224)

Time: 132947.966 ms (02:12.948)
 total_rows 
------------
  100000000
(1 row)

Time: 34617.641 ms (00:34.618)

Time: 222068.433 ms (03:42.068)
 total_rows 
------------
  100000000
(1 row)

Time: 1213.947 ms (00:01.214)
```
## Related Issues

Closes #<issue-number> or links to related issues.

## Changes

- 
- 
- 

## Checklist

- [ ] Code builds correctly
- [ ] Tests have been added or updated
- [ ] Documentation updated if necessary
- [ ] I have reviewed my own changes